# Add migration tool and use protos in Bolt store