--- a/graph/mongo/quadstore.go
+++ b/graph/mongo/quadstore.go
@@ -34,7 +34,13 @@ const DefaultDBName = "cayley"
 const QuadStoreType = "mongo"
 
 func init() {
-	graph.RegisterQuadStore(QuadStoreType, true, newQuadStore, createNewMongoGraph, nil)
+	graph.RegisterQuadStore(QuadStoreType, graph.QuadStoreRegistration{
+		NewFunc:           newQuadStore,
+		NewForRequestFunc: nil,
+		UpgradeFunc:       nil,
+		InitFunc:          createNewMongoGraph,
+		IsPersistent:      true,
+	})
 }
 
 var (