--- a/graph/proto/serializations.proto
+++ b/graph/proto/serializations.proto
@@ -1,0 +1,43 @@
+// Copyright 2015 The Cayley Authors. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+syntax = "proto3";
+
+package proto;
+
+message LogDelta {
+  uint64 ID = 1;
+  Quad Quad = 2;
+  int32 Action = 3;
+  uint64 Timestamp = 4;
+}
+
+message HistoryEntry {
+  repeated uint64 History = 1;
+}
+
+message NodeData {
+  string Name = 1;
+  int64 Size = 2;
+}
+
+message Quad {
+  string subject = 1;
+  string predicate = 2;
+  string object = 3;
+  string label = 4;
+}
+
+
+