Fixes #227 

So JSON serialization has a pretty sizeable overhead. By converting the bolt data store to using Protobufs instead of JSON, there's a pretty sizeable speedup for queries on Bolt:

```
$ benchx boltnoproto boltwproto
benchmark                                   old ns/op     new ns/op     delta	mult
BenchmarkNamePredicate                      536192        519102        -3.19%	-1.03x
BenchmarkLargeSetsNoIntersection            25406759      16517821      -34.99%	-1.54x
BenchmarkVeryLargeSetsSmallIntersection     55738157      38773174      -30.44%	-1.44x
BenchmarkHelplessContainsChecker            72689377      50568672      -30.43%	-1.44x
BenchmarkHelplessNotContainsFilms           125709712     82282494      -34.55%	-1.53x
BenchmarkHelplessNotContainsActors          101334859     67082261      -33.80%	-1.51x
BenchmarkNetAndSpeed                        5530942       4847584       -12.36%	-1.14x
BenchmarkKeanuAndNet                        4399839       4081278       -7.24%	-1.08x
BenchmarkKeanuAndSpeed                      5144942       4554785       -11.47%	-1.13x
BenchmarkKeanuOther                         16995806      12897262      -24.12%	-1.32x
BenchmarkKeanuBullockOther                  26998991      19858152      -26.45%	-1.36x
```

This means that data in the data file needs to migrate. So, I added the plumbing and rules by which migrations can happen. This is a general framework moving forward that will allow us to change the data format without forcing everybody to reload their data (the movie db takes me about a minute to load, but about 20 seconds to convert)