@Quentin-M 2015-06-24
**Benefits**: 
- Reduce the mean time of my requests from 324ms to 111ms (-65.74%)

**Drawbacks**: 
- The migration tool's memory consumption grown up to 25GB while converting a 13.96GB database (13,094,491 quads). I believe it has been fixed by 41f5573 thought.
- The new database size is 26.84GB (+92%)