@barakmich 2015-10-05
Thanks to @Quentin-M's review (and stress testing) this is going in. Things'll be faster, and the auto-update for smaller dbs shouldn't be a problem. (There's also the tool)